vioinput: Initial vioinput commit
This commit adds a new driver for the virtio input devices
with VEN 1AF4 and DEV 1052. The initial implementation exposes
all input devices as a basic 3-button mouse so it's useful
only for the mouse device (-device virtio-mouse-pci) for now.

TODO:
- build HID descriptors dynamically depending on the virtual
device capabilities
- implement status queue support (keyboard LEDs and such)

The driver is a miniport HID driver, acting like a lower
filter for MsHidKmdf.sys which ships with Windows 7 and newer.