virtiolib: Add virtio_query_queue_allocation
Some drivers may not be able to perform memory allocations on
demand at an arbitrary time. Instead, they need to be able reserve
both non-paged pool and physically-contiguous regions upfront.
The new function allows the driver to query the amount of both
types of memory needed to create a given virt queue. It is
analogous to VirtIODeviceQueryQueueAllocation in the old API set.