virtiolib: Drop run-time queue name support
It is hard to imagine a situation when seeing the queue name as
a string in the debugger would be of any particular use. There is
a simple and unique mapping from queue indices to queue names for
each virtio device.

This commit deletes the name field from queue data structures
along with all the supporting code.